Исправление ошибок в позиционировании, мелкие стилистические правки

+++ index.php
@@ -11,8 +11,8 @@
 		<script type="text/javascript" src="/starter.js"></script>
 		<link rel="stylesheet" type="text/css" href="styles.css"></link>
 		<script type="text/javascript">
-			var position_X = <?= (int)$_GET['x']; ?>;
-			var position_Y = <?= (int)$_GET['y']; ?>;
+			var position_X = <?= (float)$_GET['x']; ?>;
+			var position_Y = <?= (float)$_GET['y']; ?>;
 			var scale = 100;
 		</script>
  	</head>
